Add title blog index component
