+++ src/components/title-blog-index/index.js
@@ -1,0 +1,1 @@
+export { TitleBlogIndex } from "./title-blog-index";
